- Added GRPC code to get user from the userinfo from the userservice to the chirpservice - Fixed a few issues, user queries now fetches username too

+++ internal/auth/claims.go
@@ -3,6 +3,6 @@ package auth
 import "github.com/golang-jwt/jwt/v5"
 
 type CustomClaim struct {
-	UserID int64 `json:"user_id"`
+	UserID int `json:"user_id"`
 	jwt.RegisteredClaims
 }
